ClientQUIC/pyproject.toml
[tool.poetry]
name = "clientquic"
version = "0.1.0"
description = ""
authors = ["uibo <boui000210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
eth-utils = "^5.3.0"
pycryptodome = "^3.23.0"
rlp = "^4.1.0"
aioquic = "^1.2.0"
cryptography = "^45.0.3"
coincurve = "^21.0.0"
eciespy = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
